Added documentation, fixed minor bugs
