--- src/main/java/BotException.java
@@ -1,6 +1,0 @@
-public class BotException extends Exception{
-    public BotException(String msg) {
-        super(msg);
-    }
-}
-
