--- src/main/java/EmptyListException.java
@@ -1,5 +1,0 @@
-public class EmptyListException extends BotException{
-    public EmptyListException() {
-        super("Oops, there are no tasks in your list!");
-    }
-}
