--- src/main/java/InvalidTaskException.java
@@ -1,5 +1,0 @@
-public class InvalidTaskException extends BotException{
-    public InvalidTaskException(String msg) {
-        super(msg);
-    }
-}
