+++ src/main/java/exceptions/BotException.java
@@ -1,0 +1,16 @@
+package exceptions;
+
+/**
+ * Supertype of all exceptions that occur specific to the bot.
+ */
+public class BotException extends Exception {
+    /**
+     * Default constructor.
+     *
+     * @param msg Message to be displayed when getMessage is called.
+     */
+    public BotException(String msg) {
+        super(msg);
+    }
+}
+
